+++ src/views/File.vue
@@ -46,22 +46,29 @@ export default {
     async getBreadcrumbs(folder) {
       if (!folder) return;
 
-      const breadcrumbs = [];
-
-      let firstBreadcrumbText;
-      if (this.currentFile.appID === "drive") {
-        firstBreadcrumbText = this.isFileOwner()
-          ? this.$t("pageHeaders.MyDrive")
-          : this.$t("pageHeaders.SharedWithMe");
-      } else {
+      const hierarchy = folder.id
+        ? await filesApi.getFolderHierarchy(folder.id)
+        : [];
+      const getExternalNetworkFirstBreadcrumb = (appID) => {
         const externalNetworkDest = this.externalNetworkDests.find(
-          (externalNetworkDest) =>
-            this.currentFile.appID === externalNetworkDest.appID
+          (externalNetworkDest) => appID === externalNetworkDest.appID
         );
-        firstBreadcrumbText = this.$t("pageHeaders.ExternalTransferred", {
+        return this.$t("pageHeaders.ExternalTransferred", {
           networkName: externalNetworkDest.label,
         });
-      }
+      };
+      const getDriveFirstBreadcrumb = (ownerId) =>
+        isOwner(ownerId)
+          ? this.$t("pageHeaders.MyDrive")
+          : this.$t("pageHeaders.SharedWithMe");
+
+      const breadcrumbs = [];
+
+      // In case there is an hierarchy, it checks the first folder. Otherwise, it checks about the current file.
+      const firstFileOrFolder = hierarchy && hierarchy.length > 0 ? hierarchy[0] : this.currentFile;
+      const firstBreadcrumbText = firstFileOrFolder.isExternal
+        ? getExternalNetworkFirstBreadcrumb(firstFileOrFolder.appID)
+        : getDriveFirstBreadcrumb(firstFileOrFolder.ownerId);
 
       breadcrumbs.push({
         value: undefined,
@@ -69,23 +76,19 @@ export default {
         disabled: false,
       });
 
-      if (folder && folder.id) {
-        const hierarchy = await filesApi.getFolderHierarchy(folder.id);
-
-        hierarchy.forEach((folder) => {
-          breadcrumbs.push({
-            value: folder,
-            text: folder.name,
-            disabled: false,
-          });
-        });
-
+      hierarchy.forEach((folder) => {
         breadcrumbs.push({
           value: folder,
           text: folder.name,
           disabled: false,
         });
-      }
+      });
+
+      breadcrumbs.push({
+        value: folder,
+        text: folder.name,
+        disabled: false,
+      });
 
       this.breadcrumbs = breadcrumbs;
     },
@@ -101,9 +104,6 @@ export default {
       } else {
         this.$router.push({ path: "/folders", query: { id: folder.id } });
       }
-    },
-    isFileOwner() {
-      return isOwner(this.currentFile.ownerId);
     },
   },
 };